feat(point): 구현체 분리
