+++ src/main/java/com/popcorntalk/domain/point/service/PointRecordService.java
@@ -1,23 +1,15 @@
 package com.popcorntalk.domain.point.service;
 
-import com.popcorntalk.domain.point.entity.PointRecord;
-import com.popcorntalk.domain.point.repository.PointRecordRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+public interface PointRecordService {
 
-@Service
-@RequiredArgsConstructor
-public class PointRecordService {
-
-    private final PointRecordRepository pointRecordRepository;
-
-    @Transactional
-    public void createPointRecord(Long pointId, int previousPoint, int amount, int finalPoint) {
-
-        PointRecord pointRecord = PointRecord.createOf(
-            pointId, previousPoint, amount, finalPoint
-        );
-        pointRecordRepository.save(pointRecord);
-    }
+    /**
+     * 포인트 기록 생성 메서드
+     * 포인트 생성 및 DB 저장
+     *
+     * @param pointId 포인트 Id
+     * @param previousPoint 이전 포인트
+     * @param amount 포인트
+     * @param finalPoint 이후 포인트
+     */
+    void createPointRecord(Long pointId, int previousPoint, int amount, int finalPoint);
 }
