+++ src/main/java/com/popcorntalk/domain/point/service/PointRecordServicempl.java
@@ -1,0 +1,24 @@
+package com.popcorntalk.domain.point.service;
+
+import com.popcorntalk.domain.point.entity.PointRecord;
+import com.popcorntalk.domain.point.repository.PointRecordRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class PointRecordServicempl implements PointRecordService {
+
+    private final PointRecordRepository pointRecordRepository;
+
+    @Override
+    @Transactional
+    public void createPointRecord(Long pointId, int previousPoint, int amount, int finalPoint) {
+
+        PointRecord pointRecord = PointRecord.createOf(
+            pointId, previousPoint, amount, finalPoint
+        );
+        pointRecordRepository.save(pointRecord);
+    }
+}
